feature: adds option to prefetch all images when loading a card pool
Images are currently downloaded on-demand on mouseover, which can feel
sluggish. This change adds a new preference that, when enabled, will
prefetch all images from a (non-infinite) card pool when opening it.
Primarily for all Sanction Format modes.

Disabled by default.

Added only to desktop. Although the mobile version also loads
images on-demand, it does so as soon as the cards are visible, which
is less of an issue.

Note: will require adjustment when `imageKey` refactor (#7757) lands.